+++ data/list.py
@@ -16,6 +16,7 @@ logger = logging.getLogger(__name__)
 firestore_client = get_firestore_client()
 
 def _get_list_ref(chat_id: str, list_name: str):
+    logger.debug(f"▶️ [DEBUG] Getting list reference for chat_id: {chat_id}, list_name: {list_name}")
     return firestore_client \
         .collection("users") \
         .document(chat_id) \
@@ -24,6 +25,7 @@ def _get_list_ref(chat_id: str, list_name: str):
 
 
 def _get_items_ref(chat_id: str, list_name: str):
+    logger.debug(f"▶️ [DEBUG] Getting items reference for chat_id: {chat_id}, list_name: {list_name}")
     return _get_list_ref(chat_id, list_name) \
         .collection("items")
 
@@ -42,6 +44,7 @@ def remove_item_from_list(chat_id: str, list_name: str, item_id: str):
     return _get_items_ref(chat_id, list_name).document(item_id).delete()
 
 
-def get_list(chat_id: str, list_name: str) -> list[str]:
+def get_list(chat_id: str, list_name: str) -> list[dict]:
+    logger.debug(f"▶️ [DEBUG] Fetching list: {list_name} for chat_id: {chat_id}")
     docs = list(_get_items_ref(chat_id, list_name).stream())
-    return [doc.to_dict()["text"] for doc in docs]
+    return [{"id": doc.id, "text": doc.to_dict()["text"]} for doc in docs]
